Add Focus on new code
